scripts/requirements.txt
httpx
weaviate-client>=4.0
pypdf
